deploy models container to main (#30)
* Create __init__.py

* Delete bambooai/models.py

* Update __init__.py

* Squashed commit of the following:

commit 3f2927d86e9e7b26a3d4cfb116b3198ba617ec49
Author: Ranu Yulianto <rn.yulianto@gmail.com>
Date:   Tue May 20 15:28:07 2025 +0700

    Move models to dedicated folder

commit 6ab990261dbb8f1ba5262e836e7bc97812d225a0
Merge: a949990 39f7881
Author: Ranu Yulianto <rn.yulianto@gmail.com>
Date:   Mon May 19 21:20:56 2025 +0700

    Merge branch 'main' into ref_import

commit a9499909fa42e409f332b23482b797a5b4a8e5f1
Author: Ranu Yulianto <rn.yulianto@gmail.com>
Date:   Mon May 19 15:45:09 2025 +0700

    Fixed import in web_app

commit ff76fa811ea644bb3cd5163c3efa75b5334ae6df
Author: Ranu Yulianto <rn.yulianto@gmail.com>
Date:   Sun May 18 09:07:01 2025 +0700

    Initiate tests using pytest

commit b24e32712ff207c46ee60d4150306fc28e516a56
Author: Ranu Yulianto <rn.yulianto@gmail.com>
Date:   Sun May 18 09:06:28 2025 +0700

    Changed import style to absolute
