+++ bambooai/models/gemini_models_openai.py
